requirements.txt
streamlit==1.26.0
Pillow==9.5.0
PyQRCode==1.2.1